etl/requirements.txt
pandas==1.5.3
psycopg2-binary==2.9.9
pytest==7.4.4
pytest-mock==3.12.0
pytest-postgresql==5.0.0
python-dotenv==1.0.0
